requirements.txt
fredapi
yfinance
python-dotenv

# ML / Deep Learning
torch

# Testing
pytest

# Scheduling (optional — for daemon mode)
# apscheduler
